% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cuci_description.R
\name{add_cuci_description}
\alias{add_cuci_description}
\title{Add CUCI (SITC) level descriptions based on Comex Stat's CUCI table (STIC table)}
\usage{
add_cuci_description(
  x,
  level = c("basic_heading", "subgroup", "group", "division", "section", "all"),
  drop_key = TRUE
)
}
\description{
Add CUCI (SITC) level descriptions based on Comex Stat's CUCI table (STIC table)
}
