% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_isic_description.R
\name{add_isic_description}
\alias{add_isic_description}
\title{Add ISIC level descriptions based on Comex Stat's ISIC table}
\usage{
add_isic_description(
  x,
  lang = c("en", "pt", "es"),
  level = c("class", "group", "division", "section", "all"),
  drop_key = TRUE
)
}
\description{
Add ISIC level descriptions based on Comex Stat's ISIC table
}
