% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ncm_description.R
\name{add_ncm_description}
\alias{add_ncm_description}
\title{Add Comex Stat's NCM descriptions}
\usage{
add_ncm_description(x, lang = c("en", "pt", "es"), drop_key = FALSE)
}
\description{
Add Comex Stat's NCM descriptions
}
