% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_state_name.R
\name{add_state_name}
\alias{add_state_name}
\title{Add state name based on Comex Stat's state code}
\usage{
add_state_name(x, drop_key = TRUE)
}
\description{
Add state name based on Comex Stat's state code
}
