% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bec_table}
\alias{bec_table}
\title{Comex Stat's Broad Economic Category (BEC) codes and descriptions table}
\format{
A tibble with 19 rows and
12 variables:
\describe{
\item{bec_n3_code}{BEC n3 code.}
\item{bec_n3_desc_pt}{Description of BEC n3 in Portuguese.}
\item{bec_n3_desc}{Description of BEC n3 in English.}
\item{bec_n3_desc_es}{Description of BEC n3 in Spanish.}
\item{bec_n2_code}{BEC n2 code.}
\item{bec_n2_desc_pt}{Description of BEC n2 in Portuguese.}
\item{bec_n2_desc}{Description of BEC n2 in English .}
\item{bec_n2_desc_es}{Description of BEC n2 in Spanish.}
\item{bec_n1_code}{BEC n1 code.}
\item{bec_n1_desc_pt}{Description of BEC n1 in Portuguese.}
\item{bec_n1_desc}{Description of BEC n1 in English.}
\item{bec_n1_desc_es}{Description of BEC n1 in Spanish.}
}
}
\source{
\url{https://balanca.economia.gov.br/balanca/bd/tabelas/NCM_CGCE.csv}
}
\usage{
bec_table
}
\description{
A table containing BEC codes and descriptions.
}
\examples{
data(bec_table)
head(bec_table)
}
\keyword{datasets}
