% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country_table}
\alias{country_table}
\title{Comex Stat's Country Table}
\format{
A tibble with 281 rows and
6 variables:
\describe{
\item{country_code}{Brazilian country code (character)}
\item{country_code_ison3}{ISON3 country code (character)}
\item{country_code_isoa3}{ISOA3 country code (character)}
\item{country_name_pt)}{Country name portuguese (character)}
\item{country_name}{Country name (character)}
\item{country_name_es}{Country name spanish (character)}
}
}
\source{
\url{https://balanca.economia.gov.br/balanca/bd/tabelas/PAIS.csv}
}
\usage{
country_table
}
\description{
A table containing country codes and names
}
\examples{
data(country_table)
head(country_table)
}
\keyword{datasets}
