% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\docType{data}
\name{cuci_table}
\alias{cuci_table}
\title{Comex Stat's CUCI (STIC) code and description table - only in portuguese}
\format{
A tibble with 2971 rows and 6 columns:
\describe{
\item{cuci_basic_heading_code}{CUCI basic heading code}
\item{cuci_basic_heading_desc_pt}{CUCI basic heading description in Portuguese.}
\item{cuci_subgroup_desc_pt}{CUCI subgroup description in Portuguese.}
\item{cuci_group_desc_pt}{CUCI group description in Portuguese.}
\item{cuci_division_desc_pt}{CUCI division description in Portuguese.}
\item{cuci_section_desc_pt}{CUCI section description in Portuguese.}
}
}
\source{
\url{https://balanca.economia.gov.br/balanca/bd/tabelas/NCM_CUCI.csv}
}
\description{
A table containing unit codes and descriptions
}
\keyword{internal}
