% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_states.R
\name{filter_states}
\alias{filter_states}
\title{Wrapper to Filter Brazilian States}
\usage{
filter_states(x)
}
\arguments{
\item{x}{A data frame that must include a column named \code{state}.}
}
\value{
A filtered data frame containing only rows with valid Brazilian
states in the \code{state} column.
}
\description{
A wrapper around \code{\link[dplyr:filter]{dplyr::filter()}} to retain only rows with
Brazilian state abbreviations. The column state also has non-state data.
So this wrapper filter only state data.
}
\details{
The \code{state} column must exist in the input data frame. If missing,
the function throws an error. The internal list of valid states is defined
in the object \code{brazilian_states}, assumed to be loaded with the package.
}
\examples{
df <- data.frame(state = c("SP", "RJ", "ND"), value = 1:3)
filter_states(df)

}
