% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isic_table}
\alias{isic_table}
\title{Comex Stat's International Standard Industrial Classification (ISIC) category
codes and descriptions table}
\format{
A tibble with 420 rows and
16 variables:
\describe{
\item{isic_class_code}{ISIC class code.}
\item{isic_class_desc_pt}{Description of ISIC class in Portuguese.}
\item{isic_class_desc}{Description of ISIC class in English.}
\item{isic_class_desc_es}{Description of ISIC class in Spanish.}
\item{isic_group_code}{Code of the ISIC group.}
\item{isic_group_desc_pt}{Description of ISIC group in Portuguese.}
\item{isic_group_desc}{Description of ISIC group in English .}
\item{isic_group_desc_es}{Description of ISIC group in Spanish.}
\item{isic_division_code}{ISIC division code.}
\item{isic_division_desc_pt}{Description of ISIC division in Portuguese.}
\item{isic_division_desc}{Description of ISIC division in English.}
\item{isic_division_desc_es}{Description of ISIC division in Spanish.}
\item{isic_section_code}{ISIC section code.}
\item{isic_section_desc_pt}{Description of ISIC section in Portuguese.}
\item{isic_section_desc}{Description of ISIC section in English.}
\item{isic_section_desc_es}{Description of ISIC section in Spanish.}
}
}
\source{
\url{https://balanca.economia.gov.br/balanca/bd/tabelas/NCM_ISIC.csv}
}
\usage{
isic_table
}
\description{
A table containing ISIC category codes and descriptions.
}
\examples{
data(isic_table)
head(isic_table)
}
\keyword{datasets}
