% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncm_table}
\alias{ncm_table}
\title{Comex Stat's NCM Mapping Table}
\format{
A tibble with 13,724 rows and
9 variables:
\describe{
\item{ncm}{NCM Product code}
\item{unit_code}{Unit code for NCM}
\item{cuci_basic_heading_code}{STIC basic heading code}
\item{bec_n3_code)}{Broad economic category code}
\item{co_siit}{SIIT code}
\item{isic_class_code}{ISIC class code}
\item{ncm_description}{Product description}
\item{ncm_description_pt}{Product description portuguese}
\item{ncm_description_es}{Product description spanish}
}
}
\source{
\url{https://balanca.economia.gov.br/balanca/bd/tabelas/NCM.csv}
}
\usage{
ncm_table
}
\description{
A table containing ncm and other product category correlations.
}
\examples{
data(ncm_table)
head(ncm_table)
}
\keyword{datasets}
