% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_table}
\alias{state_table}
\title{Comex Stat's State Table}
\format{
A tibble with 34 rows and
2 variables::
\describe{
\item{state}{State abbreviation (character)}
\item{state_name}{Full state name (character)}
}
}
\source{
\url{https://balanca.economia.gov.br/balanca/bd/tabelas/UF.csv}
}
\usage{
state_table
}
\description{
A table containing state abbreviations and full names.
}
\examples{
data(state_table)
head(state_table)
}
\keyword{datasets}
