% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cs_db.R
\name{update_cs_db}
\alias{update_cs_db}
\title{Update Local Comex Stat Database}
\usage{
update_cs_db(dest_dir, start_year = NULL, timeout = 2000)
}
\arguments{
\item{dest_dir}{Character string. Path to the base directory of the local database (same used in \code{create_cs_db()}).}

\item{start_year}{Integer. If supplied, besides updating, the function will check if the desired first year
is included in the Comex Stat dataset. In case it is not it will download all
years not included in the dataset starting from start_year. .
If \code{NULL}, the function will only will try to update the dataset.}

\item{timeout}{Integer. Timeout limit in seconds for HTTP requests. Default is \code{2000}.}
}
\value{
No return value. Side effect: updates existing local Arrow datasets with newer data from the official Comex Stat raw files.
}
\description{
Checks if the local database is outdated and updates data if necessary.
}
\details{
This function compares the date in the local Arrow datasets with the most recent date in
the official database files. In case local dataset is not updated, the function
will update local data.

If the local database is already up-to-date, the function stops with a message.
}
\examples{
\dontrun{
update_cs_db(dest_dir = "data/comexstat")
}

}
\seealso{
\code{\link[=create_cs_db]{create_cs_db()}}
}
