% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_data.R
\docType{data}
\name{cuci_table}
\alias{cuci_table}
\title{Comex Stat's CUCI code and description table - only in portuguese}
\format{
A tibble with 2971 rows and 6 columns:
\describe{
\item{no_cuci_item}{CUCI item code}
\item{no_cuci_item}{CUCI item description in Portuguese.}
\item{no_cuci_sub}{CUCI subgroup description in Portuguese.}
\item{no_cuci_grupo}{CUCI group description in Portuguese.}
\item{no_cuci_divisao}{CUCI division description in Portuguese.}
\item{no_cuci_sec}{CUCI section description in Portuguese.}
}
}
\source{
\url{https://balanca.economia.gov.br/balanca/bd/tabelas/NCM_CUCI.csv}
}
\description{
A table containing unit codes and descriptions
}
\keyword{internal}
